fix: вынести проверки в сервис

+++ src/main/java/ru/practicum/shareit/user/UserServiceImpl.java
@@ -3,6 +3,8 @@ package ru.practicum.shareit.user;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.practicum.shareit.exception.ConflictException;
+import ru.practicum.shareit.exception.UserNotFoundException;
 
 import java.util.List;
 
@@ -14,12 +16,34 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User addUser(User user) {
-        return userRepository.addUser(user);
+        if (userRepository.duplicateEmail(user.getEmail())) {
+            throw new ConflictException("Пользователь с почтой: " + user.getEmail() + " уже существует.");
+        } else {
+            return userRepository.addUser(user);
+        }
     }
+
 
     @Override
     public User updateUser(Long userId, User user) {
-        return userRepository.updateUser(userId, user);
+        User existingUser = getUserById(userId);
+        if (user.getName() != null) {
+            existingUser.setName(user.getName());
+            log.info("Редактирование имени: {}", user.getName());
+        }
+        if (user.getEmail() != null) {
+            String newEmail = user.getEmail();
+            if (!newEmail.equals(existingUser.getEmail())) {
+                if (!userRepository.duplicateEmail(newEmail)) {
+                    existingUser.setEmail(newEmail);
+                    log.info("Редактирование почты: {}", user.getEmail());
+                } else {
+                    throw new ConflictException(
+                        "Пользователь с почтой: " + user.getEmail() + " уже существует.");
+                }
+            }
+        }
+        return userRepository.updateUser(userId, existingUser);
     }
 
     @Override
@@ -29,12 +53,21 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User getUserById(Long userId) {
-        return userRepository.getUserById(userId);
+        if (userRepository.containsUser(userId)) {
+            return userRepository.getUserById(userId);
+        } else {
+            throw new UserNotFoundException("Пользователь с id: " + userId);
+        }
     }
 
     @Override
     public void deleteUserById(Long userId) {
-        userRepository.deleteUserById(userId);
+        if (userRepository.containsUser(userId)) {
+            userRepository.deleteUserById(userId);
+        } else {
+            throw new UserNotFoundException("Пользователь с id: " + userId);
+        }
+
     }
 
 }
